requirements.txt
selenium
webdriver_manager
python-dotenv
APScheduler
flask
flask-cors
pillow